working gpu version with bugs

+++ gpu/src/kernels.cu
@@ -17,7 +17,7 @@ __global__ void fillInitialCondition(REAL* a, REAL* F, REAL *G, size_t l, size_t
 
 	a[(l)*M*N*O + (gid_x)*N*O + (gid_y)*O + gid_z] = a_0[gid_x];
 	F[(l)*M*N*O + (gid_x)*N*O + (gid_y)*O + gid_z] = q*(dtheta*gid_y);
-	G[(l)*M*N*O + (gid_x)*N*O + (gid_y)*O + gid_z] = p*((dt*l)/L - dphi*gid_z);
+	G[(l)*M*N*O + (gid_x)*N*O + (gid_y)*O + gid_z] = p*((dt*(REAL)l)/(REAL)L - dphi*(REAL)gid_z);
 }
